Removed NumericMatrix from other .cpp files

+++ src/affinealignobj.h
@@ -17,21 +17,21 @@ std::vector<char> EnumToChar(std::vector<TracebackType> v);
 
 struct AffineAlignObj
 {
-  std::vector<float> M; // Match or Mismatch matrix, residues of A and B are aligned without a gap. M(i,j) = Best score upto (i,j) given Ai is aligned to Bj.
-  std::vector<float> A; // Insert in sequence A, residue in A is aligned to gap in B. A(i,j) is the best score given that Ai is aligned to a gap in B.
-  std::vector<float> B; // Insert in sequence B, residue in B is aligned to gap in A. B(i,j) is the best score given that Bj is aligned to a gap in A.
+  std::vector<double> M; // Match or Mismatch matrix, residues of A and B are aligned without a gap. M(i,j) = Best score upto (i,j) given Ai is aligned to Bj.
+  std::vector<double> A; // Insert in sequence A, residue in A is aligned to gap in B. A(i,j) is the best score given that Ai is aligned to a gap in B.
+  std::vector<double> B; // Insert in sequence B, residue in B is aligned to gap in A. B(i,j) is the best score given that Bj is aligned to a gap in A.
   std::vector<TracebackType> Traceback;
   int signalA_len; // Number of data-points in signal A
   int signalB_len; // Number of data-points in signal B
-  float GapOpen; // Penalty for Gap opening
-  float GapExten; // Penalty for Gap extension
+  double GapOpen; // Penalty for Gap opening
+  double GapExten; // Penalty for Gap extension
   // For single gap: Penalty = GapOpen
   // For two consecutive gaps: Penalty = GapOpen + GapExten
   // For n consecutive gaps: Penalty = GapOpen + (n-1)*GapExten
   bool FreeEndGaps; // True for Overlap alignment
   std::vector<int> indexA_aligned; // Aligned signalA indices after affine alignment
   std::vector<int> indexB_aligned; // Aligned signalB indices after affine alignment
-  std::vector<float> score;  // Score along the aligned path
+  std::vector<double> score;  // Score along the aligned path
 
   // Not a default constructor
   AffineAlignObj(int ROW_SIZE, int COL_SIZE)
